sun2000: rework saving influx data
Influx data was saved as a single sample when all parameters was read.
As long as the internal wifi was used for connection it was quite ok,
as reading all parameters took about 4 secs.
Using a smart dongle connection though can raise up this time even
to about 18 secs!
As the result, samples was wrongly timestamp-ed - especially for
the very early-read parameters from the set.

This commit change the way the data is saved - now it is saving
each parameter to each own influx data series and it is now saved
right after getting a response for a single parameter, thus the
timestamping is much more accurate.

This is even leading to better managable data in influxdb (splitted
into series instead one big param "table").