main: port to newest simplelog
simplelog is now using `termcolor` instead of deprecated `term`
underlying library for the color terminal.

termcolor does automatic color selection differently:
https://github.com/BurntSushi/termcolor#automatic-color-selection

This means termcolor does not fail the same way term did, but will try
to defer from the environment, if colors are supported and desired or
not. But it will never fail by default and always be able to print
something.

This commit is chaning the old console_logger creation to this new
non-failing scheme.