sun2000: give up after 3x param read error; cosmetics
When the same parameter cannot be obtained from inverter for three times,
there is no need to ask for another one; it is probably a better idea
to break and try to reconnect.

The commit also change logging in a way to treat first two read attempts
as a warning instead of error. Throw an error only when unable to obtain
a single parameter from inverter for three times.