sun2000: optimize reading by grouping registers
Groups are done automatically basing on set of registers
desired to read.

The idea was discovered by @wlcrs and implemented in:
https://github.com/wlcrs/huawei-solar-lib/blob/04ee7a765772f30f98aeca712344fd791d5579f9/src/huawei_solar/huawei_solar.py#L243
Then it was also implemented by @amreo in his fork:
https://github.com/amreo/hard-sun2000-influxdb-datalogger/commit/1ec907c16217253c216ad07c0ebe063dc31df3ba

In my environment whole reading set went from 5.5s down to 2.3s.

This commit is not yet rust-formatted due to see the relevant
changes easily.